requirements.txt
fastapi
uvicorn
pydantic
python-docx
playwright
pytest-playwright